+++ lib/src/utils.dart
@@ -1,3 +1,4 @@
 library icomoon_generator.utils;
 
 export 'utils/flutter_class_gen.dart';
+export 'utils/enum_class.dart';